+++ net/minecraft/entity/mob/CreeperEntity.java
@@ -197,7 +197,7 @@ implements SkinOverlayOwner {
     }
 
     public float getClientFuseTime(float timeDelta) {
-        return MathHelper.lerp(timeDelta, this.lastFuseTime, this.currentFuseTime) / (this.fuseTime - 2);
+        return MathHelper.lerp(timeDelta, (float)this.lastFuseTime, (float)this.currentFuseTime) / (float)(this.fuseTime - 2);
     }
 
     public int getFuseSpeed() {
